Moving public things over to core/ package
